The desktop run works from Gradle, as does the "test" target.

I still need to try packaging the Mac app, and it looks like it is still
expecting a Java 8 JRE library.

+++ core/src/main/java/com/vzome/core/exporters/PdbExporter.java
@@ -74,7 +74,7 @@ public class PdbExporter extends Exporter3d
             RealVector rv = mModel .renderVector( atom .location );
             System .out .println( atom .location .toString() );
             locations .append( String .format( "HETATM%5d He   UNK  0001     %7.3f %7.3f %7.3f\n",
-                    atom .index, (float) rv.x * scaleFactor, (float) rv.y * scaleFactor, (float) rv.z * scaleFactor ) );
+                    atom .index, rv.x * scaleFactor, rv.y * scaleFactor, rv.z * scaleFactor ) );
             neighbors .append( String .format( "CONECT%5d", atom .index ) );
             for (Atom neighbor : atom .neighbors) {
                 neighbors .append( String .format( "%5d", neighbor .index ) );
